Pull request log info stuff into `Request`.
